Add an iterator chunking extension
I regularly want to chunk some data, usually data that I know will line
up with my chunk size. To appease the type system, I need the Default
trait on the iterator item here, because it cannot statically verify
that the chunks do actually line up.

This is similar to the chunking that is already in unstable, but that
one uses a Result Error to return a smaller chunk at the end. I find
filling up with default() and then returning None more idiomatic.

One could also consider wrapping each item in an Option, but that would
be a hassle to unwrap.